Rework Rust-C ffi to allocate and deallocate with size and align.
Instead of using the Posix/C API we define our own Rust API.
This allows us to pass size and alignment information for an extra
check, "delete_size_mismatch".

This is a breaking change for scudo-sys but not for scudo hence
the changes in the semantic versions.

While I was at it, I also renamed and changed the arguments of some
Rust-C-FFI functions to make them nicer to use.

Change-Id: Id9bd00dd2d30e44cc4b30c7dd52c78326324ddab